% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getYaml}
\alias{getYaml}
\title{Get YAML Markup (generic)}
\usage{
getYaml(from, ctx = NULL, ...)
}
\arguments{
\item{from}{\strong{Signature argument}.
Object containing YAML markup.}

\item{ctx}{\strong{Signature argument}.
Markup proecessing context.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\description{
Identifies lines that correspond to YAML markup.
}
\examples{
\dontrun{

from <- function() {
  "reactive-ref: {id: x_1, where:.GlobalEnv, as: ref_1}"
}
res <- getYaml(
  from = from,
  ctx = ReactiveReferenceYamlContext.S3()
)
res
res$yaml
res$index
res$src

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/fromyaml}
}
\seealso{
\code{
  	\link[fromyaml]{getYaml-function-ReactiveReferenceYamlContext.S3-method}
}
}

