% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{getYamlFromComments,function,YamlContext.ObjectReference.S3-method}
\alias{getYamlFromComments,function,YamlContext.ObjectReference.S3-method}
\alias{getYamlFromComments-function-YamlContext.ObjectReference.S3-method}
\title{Get YAML Markup From Comments (function-YamlContext.ObjectReference.S3)}
\usage{
\S4method{getYamlFromComments}{`function`,YamlContext.ObjectReference.S3}(from,
  ctx = NULL, where = parent.frame(), ...)
}
\arguments{
\item{from}{\code{\link{missing}}.}

\item{ctx}{\strong{Signature argument}.
Markup proecessing context.}

\item{where}{\code{\link{environment}}.
Environment in which to assign the function in the \code{src} field of
class \code{\link[yamlr]{ObjectReferenceYaml.S3}}. Only relevant
in case the YAML has been provided via comments instead of an
inline string as this involves some additional transformation steps.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
\code{\link{character}}. Identified YAML markup.
}
\description{
See generic: \code{\link[yamlr]{getYamlFromComments}}
}
\examples{
\dontrun{

yaml <- getYamlFromComments(
  from = function() {
    ## object-ref: {id: x_1, where:.GlobalEnv, as: ref_1}
  },
  ctx = YamlContext.ObjectReference.S3()
)
yaml$original
yaml$src

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/yamlr}
}
\seealso{
\code{
   \link[yamlr]{getYamlFromComments},
    \link[yamlr]{getYaml}
}
}

