% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getYamlFromComments}
\alias{getYamlFromComments}
\title{Get YAML Markup From Comments (generic)}
\usage{
getYamlFromComments(from, ctx = NULL, where = parent.frame(), ...)
}
\arguments{
\item{from}{\strong{Signature argument}.
Object containing YAML markup.}

\item{ctx}{\strong{Signature argument}.
Markup proecessing context.}

\item{where}{\code{\link{environment}}.
Environment in which to assign the function in the \code{src} field of
class \code{\link[yamlr]{ObjectReferenceYaml.S3}}. Only relevant
in case the YAML has been provided via comments instead of an
inline string as this involves some additional transformation steps.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\description{
Identifies lines that correspond to YAML markup in comments.
}
\section{YAML syntax}{


Note that the YAML statement for specifying object references must
be as follows:

\code{object-ref: {id: {id}, where: {where}, as: {id-ref}}}

Note that elements \code{where} and \code{as} are optional and that spaces
are required between each syntax component. While the lack of spaces will
cause a parsing error in \code{\link[yaml]{yaml.load}}, additional spaces
do not casue any trouble.

Example for valid YAML statements

\itemize{
   \item{\code{object-ref: {id: x_1, where: .GlobalEnv, as: ref_1}}}
   \item{\code{object-ref: {id: x_1, where: .GlobalEnv}}}
   \item{\code{object-ref: {id: x_1}}}
   \item{\code{object-ref:  {  id:   x_1, where  : .GlobalEnv, as: ref_1  }}}
}

See the official \href{YAML website}{http://www.yaml.org/} for more
details about YAML
}
\examples{
\dontrun{

yaml <- getYamlFromComments(
  from = function() {
    ## object-ref: {id: x_1, where:.GlobalEnv, as: ref_1}
  },
  ctx = YamlContext.ObjectReference.S3()
)
yaml$original
yaml$src

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/yamlr}
}
\seealso{
\code{
  	\link[yamlr]{getYamlFromComments-function-YamlContext.ObjectReference.S3-method},
    \link[yamlr]{getYaml}
}
}

