% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{parseYaml,ObjectReferenceYaml.S3-method}
\alias{parseYaml,ObjectReferenceYaml.S3-method}
\alias{parseYaml-ObjectReferenceYaml.S3-method}
\title{Parse YAML Markup (ObjectReferenceYaml.S3)}
\usage{
\S4method{parseYaml}{ObjectReferenceYaml.S3}(yaml, strict = FALSE, ...)
}
\arguments{
\item{yaml}{\code{\link{ObjectReferenceYaml.S3}}.}

\item{strict}{\code{\link{logical}}.
\code{TRUE}: error if no YAML markup could be identified (which in turn
results in field \code{original} in class
\code{\link[yamlr]{ObjectReferencedYaml.S3}} being empty);
\code{FALSE}: no error if no YAML markup could be identified.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
\code{\link{ObjectReferenceYamlParsed.S3}}.
   Parsed YAML markup.
}
\description{
See generic: \code{\link[yamlyaml]{parseYaml}}
}
\examples{
\dontrun{

##------------------------------------------------------------------------------
## From function //
##------------------------------------------------------------------------------

## Get //
yaml <- getYaml(
  from = function() {
    "object-ref: {id: x_1, where: .GlobalEnv, as: ref_1}"
  },
  ctx = YamlContext.ObjectReference.S3()
)

## Parse //
yaml <- parseYaml(yaml = yaml)

## Inspect //
yaml
yaml$original
yaml$parsed
yaml$index
yaml$src

## Condition handling //
yaml <- getYaml(
  from = function() {
    "Invalid YAML markup"
  },
  ctx = YamlContext.ObjectReference.S3()
)
yaml$original

try(getYaml(
  from = function() {
    "Invalid YAML markup"
  },
  ctx = YamlContext.ObjectReference.S3(),
  strict = TRUE
))

##------------------------------------------------------------------------------
## From expression //
##------------------------------------------------------------------------------

yaml <- getYaml(
  from = captureExpression({
    "object-ref: {id: x_1, where: .GlobalEnv, as: ref_1}"
    ref_1
  }),
  ctx = YamlContext.ObjectReference.S3()
)
yaml <- parseYaml(yaml = yaml)

## Inspect //
yaml
yaml$original
yaml$parsed
yaml$index
yaml$src

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/yamlr}
}
\seealso{
\code{
   \link[yamlyaml]{parseYaml}
}
}

