% Generated by roxygen2 (4.0.2): do not edit by hand
\name{parseYaml}
\alias{parseYaml}
\title{Parse YAML Markup (generic)}
\usage{
parseYaml(yaml, ...)
}
\arguments{
\item{yaml}{\strong{Signature argument}.
Object containing identified YAML markup as returned by
\code{\link[fromyaml]{getYaml}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\description{
Parses YAML markup as identified by \code{\link[fromyaml]{getYaml}}.
}
\examples{
\dontrun{

## Get //
yaml <- getYaml(
  from = function() {
  "reactive-ref: {id: x_1, where: .GlobalEnv, as: ref_1}"
  },
  ctx = ReactiveReferenceYamlContext.S3()
)

## Parse //
yaml <- parseYaml(yaml = yaml)

## Inspeact //
yaml
yaml$yaml
yaml$yaml_parsed
yaml$index
yaml$src

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/fromyaml}
}
\seealso{
\code{
  	\link[fromyaml]{parseYaml-function-ReactiveReferenceYamlContext.S3-method}
}
}

