% Generated by roxygen2 (4.0.2): do not edit by hand
\name{processYaml}
\alias{processYaml}
\title{Process YAML Markup (generic)}
\usage{
processYaml(from, ctx = NULL, where = parent.frame(), ...)
}
\arguments{
\item{from}{\strong{Signature argument}.
Object containing YAML markup.}

\item{ctx}{\strong{Signature argument}.
Markup proecessing context.}

\item{where}{\code{\link{environment}}.
Environment in which to assign the function in the \code{src} field of
class \code{\link[fromyaml]{ReactiveReferenceYaml.S3}}. Only relevant
in case the YAML has been provided via comments instead of an
inline string as this involves some additional transformation steps.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\description{
Convenience wrapper. Processes YAML markup in the sense that the complete
workflow associated to a given application context is executed.
}
\details{
Currently, this means that all processing steps associated with the context
\code{\link[fromyaml]{YamlContext.ReactiveReference.S3}} are executed:

\itemiz{
   \item{\code{\link[fromyaml]{getYaml}}}
   \item{\code{\link[fromyaml]{parseYaml}}}
   \item{\code{\link[fromyaml]{buildExpressionFromYaml}}}
   \item{\code{\link[fromyaml]{updateYamlSource}}}
}
}
\examples{
\dontrun{

## Get //  
yaml <- processYaml(
  from = function() {
    "reactive-ref: {id: x_1, where: .GlobalEnv, as: ref_1}"
    ref_1 * 2
  },
  ctx = YamlContext.ReactiveReference.S3()
)

## Inspect //
yaml$src
## --> note that the original YAML markup has been substituted by a 
## suitable expression for this specific context (`yaml$src$x_1$get_assign`).
## The processing is now complete and the source object can be used as is
## as it is now "self contained" in the sense that it contains all information
## that are needed in order for it to do its job.

## Actual use of the source object //
x_1 <- 10
eval(yaml$src())

## Clean up //
rm(x_1)
rm(yaml)

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/fromyaml}
}
\seealso{
\code{
  	\link[fromyaml]{processYaml-function-YamlContext.ReactiveReference.S3-method},
   \link[fromyaml]{getYaml},
   \link[fromyaml]{parseYaml},
   \link[fromyaml]{buildExpressionFromYaml},
   \link[fromyaml]{updateYamlSource}
}
}

