% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{updateYamlSource,ReactiveReferenceYamlProcessed.S3-method}
\alias{updateYamlSource,ReactiveReferenceYamlProcessed.S3-method}
\alias{updateYamlSource-ReactiveReferenceYamlProcessed.S3-method}
\title{Parse YAML Markup (ReactiveReferenceYaml.S3)}
\usage{
\S4method{updateYamlSource}{ReactiveReferenceYamlProcessed.S3}(yaml, ...)
}
\arguments{
\item{yaml}{\code{\link{ReactiveReferenceYamlProcessed.S3}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
\code{\link{ReactiveReferenceYamlProcessed}}.
 Processed YAML markup with Updated \code{src} field value.
 The original markup has been substituted by the respective expression
 (\code{$expr${ref}$get_assign} with \code{{ref}} being the ID of the
 reactive reference).
}
\description{
See generic: \code{\link[yamlyaml]{updateYamlSource}}
}
\examples{
\dontrun{

## Get //  
yaml <- getYaml(
  from = function() {
  "reactive-ref: {id: x_1, where: .GlobalEnv, as: ref_1}"
  ref_1 * 2
  },
  ctx = ReactiveReferenceYamlContext.S3()
)

## Parse //
yaml <- parseYaml(yaml = yaml)

## Build //
yaml <- buildExpressionFromYaml(yaml = yaml)

## Update //
yaml <- updateYamlSource(yaml = yaml)

## Inspect //
yaml$src
## --> note that the original YAML markup has been substituted by a 
## suitable expression for this specific context (`yaml$src$x_1$get_assign`).
## The processing is now complete and the source object can be used as is
## as it is now "self contained" in the sense that it contains all information
## that are needed in order for it to do its job.

## Actual use of the source object //
x_1 <- 10
eval(yaml$src())

## Clean up //
rm(x_1)
rm(yaml)
rm(yaml)

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/fromyaml}
}
\seealso{
\code{
   \link[yamlyaml]{updateYamlSource},
    \link[fromyaml]{getYaml},
    \link[fromyaml]{parseYaml},
    \link[fromyaml]{buildExpressionFromYaml}
}
}

