% Generated by roxygen2 (4.0.2): do not edit by hand
\name{updateYamlSource}
\alias{updateYamlSource}
\title{Update Object that originally contained YAML Markup (generic)}
\usage{
updateYamlSource(yaml, ...)
}
\arguments{
\item{yaml}{\strong{Signature argument}.
Object containing processed YAML markup as returned by
\code{\link[yamlr]{buildExpressionFromYaml}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\description{
Updates the object that originally contained YAML markup. To be more
specific, the function updates field \code{src} or class
\code{\link[yamlr]{ObjectReferenceYamlProcessed}} as returned by
\code{\link[yamlr]{buildExpressionFromYaml}}.
}
\section{YAML syntax}{


Note that the YAML statement for specifying object references must
be as follows:

\code{object-ref: {id: {id}, where: {where}, as: {id-ref}}}

Note that elements \code{where} and \code{as} are optional and that spaces
are required between each syntax component. While the lack of spaces will
cause a parsing error in \code{\link[yaml]{yaml.load}}, additional spaces
do not casue any trouble.

Example for valid YAML statements

\itemize{
   \item{\code{object-ref: {id: x_1, where: .GlobalEnv, as: ref_1}}}
   \item{\code{object-ref: {id: x_1, where: .GlobalEnv}}}
   \item{\code{object-ref: {id: x_1}}}
   \item{\code{object-ref:  {  id:   x_1, where  : .GlobalEnv, as: ref_1  }}}
}

See the official \href{YAML website}{http://www.yaml.org/} for more
details about YAML
}
\examples{
\dontrun{

## Get //  
yaml <- getYaml(
  from = function() {
  "object-ref: {id: x_1, where: .GlobalEnv, as: ref_1}"
  ref_1 * 2
  },
  ctx = YamlContext.ObjectReference.S3()
)

## Parse //
yaml <- parseYaml(yaml = yaml)

## Build //
yaml <- buildExpressionFromYaml(yaml = yaml)

## Update //
yaml <- updateYamlSource(yaml = yaml)

## Inspect //
yaml$src
## --> note that the original YAML markup has been substituted by a 
## suitable expression for this specific context (`yaml$src$x_1$get_assign`).
## The processing is now complete and the source object can be used as is
## as it is now "self contained" in the sense that it contains all information
## that are needed in order for it to do its job.

## Actual use of the source object //
x_1 <- 10
eval(yaml$src())

## Clean up //
rm(x_1)
rm(yaml)

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/yamlr}
}
\seealso{
\code{
  	\link[yamlr]{updateYamlSource-ObjectReferenceYamlProcessed.S3-method},
    \link[yamlr]{getYaml},
    \link[yamlr]{parseYaml},
    \link[yamlr]{buildExpressionFromYaml}
}
}

