% Generated by roxygen2 (4.0.2): do not edit by hand
\name{YamlContext.ObjectReference.S3}
\alias{YamlContext.ObjectReference.S3}
\title{Class: YamlContext.ObjectReference.S3}
\usage{
YamlContext.ObjectReference.S3(.x, ctx = "YamlContext.ObjectReference.S3",
  pattern = ".*object-ref:\\\\s+\\\\{\\\\s*id\\\\s*:\\\\s+\\\\w+.*\\\\}")
}
\arguments{
\item{.x}{\code{\link{ANY}}. An object of an arbitrary class whose class
   attribute should be updated so that it becomes an instance of class
   \code{YamlContext.ObjectReference.S3}. Mainly intended for rapid prototyping
   purposes}
}
\value{
Instance of class \code{YamlContext.ObjectReference.S3}.
}
\description{
Class representing the reactive reference context with respect to
processing YAML markup.
}
\section{Fields}{

\describe{
\item{\code{ctx}}{\code{\link{character}}.
Context ID. Corresponds to the class name.
Initial: \code{"YamlContext.ObjectReference.S3"}.}

\item{\code{pattern}}{\code{\link{character}}.
Pattern to recognize YAML markup.
Initial: \code{".*object-ref:\\s+\\{\\s*id\\s*:\\s+\\w+.*\\}"}.}
}}
\section{Intended use of this class}{


This S3 class, or to be more precise its constructor function, exists mainly
for rapid prototyping purposes.
This is mainly reflected in the fact, that when specifying \code{.x}, this
constructor function will simply update the \code{class} attribute of
whatever object has been provided.

However, it also allows for a more formal OOP-style of rapid
prototyping by offering explicit \emph{class fields} (all arguments except
\code{.x}). Nevertheless, it is probably advisable to switch to an
explicit formal approach such as \emph{S4} and/or \emph{Reference Classes}
once the package or application has reached a certain state of maturity.
}
\examples{
\dontrun{

## Informal use (intended mainly for rapid prototyping) //
## Takes *any* object and simply changes the class attributes
YamlContext.ObjectReference.S3(
  list(
    ctx = "abcd"
  )
)  
YamlContext.ObjectReference.S3(TRUE)  

## Formal use (explicitly using 'fields') //
res <- YamlContext.ObjectReference.S3()
ls(res)
res$ctx
res <- YamlContext.ObjectReference.S3(
  ctx = "YamlContext.ObjectReference.S3"
)
res$ctx

## Recommended: include namespace //
## Regardless if you plan on using this class in an informal or formal way
yamlr::YamlContext.ObjectReference.S3()

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/yamlr}
}

