% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{buildExpressionFromYaml,ReactiveReferenceYamlParsed.S3-method}
\alias{buildExpressionFromYaml,ReactiveReferenceYamlParsed.S3-method}
\alias{buildExpressionFromYaml-ReactiveReferenceYamlParsed.S3-method}
\title{Build Expression Based on YAML Markup (ReactiveReferenceYamlParsed.S3)}
\usage{
\S4method{buildExpressionFromYaml}{ReactiveReferenceYamlParsed.S3}(yaml, ...)
}
\arguments{
\item{yaml}{\code{\link{ReactiveReferenceYamlParsed.S3}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
\code{\link{ReactiveReferenceYamlProcessed.S3}}.
   Processed YAML markup.
}
\description{
See generic: \code{\link[yamlyaml]{buildExpressionFromYaml}}
}
\examples{
\dontrun{

## Get //  
yaml <- getYaml(
  from = function() {
  "reactive-ref: {id: x_1, where: .GlobalEnv, as: ref_1}"
  },
  ctx = ReactiveReferenceYamlContext.S3()
)

## Parse //
yaml <- parseYaml(yaml = yaml)

## Build //
res <- buildExpressionFromYaml(yaml = yaml)
res$expr$x_1
res$expr$x_1$get_assign

## Example of expression execution //
## The line that contained the YAML markup in 'from' above will be substituted
## by this expression 
x_1 <- 10
eval(res$expr$x_1$get_assign)
ref_1

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/fromyaml}
}
\seealso{
\code{
   \link[yamlyaml]{buildExpressionFromYaml}
}
}

