% Generated by roxygen2 (4.0.2): do not edit by hand
\name{buildExpressionFromYaml}
\alias{buildExpressionFromYaml}
\title{Build Expression Based on YAML Markup (generic)}
\usage{
buildExpressionFromYaml(yaml, ...)
}
\arguments{
\item{yaml}{\strong{Signature argument}.
Object containing parsed YAML markup as returned by
\code{\link[fromyaml]{parseYaml}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\description{
Builds expression(s) based on the YAML markup as identified by
\code{\link[fromyaml]{getYaml}} and parsed by \code{\link[fromyaml]{parseYaml}}.
}
\examples{
\dontrun{

## Get //  
yaml <- getYaml(
  from = function() {
  "reactive-ref: {id: x_1, where: .GlobalEnv, as: ref_1}"
  },
  ctx = ReactiveReferenceYamlContext.S3()
)

## Parse //
yaml <- parseYaml(yaml = yaml)

## Build //
yaml <- buildExpressionFromYaml(yaml = yaml)
yaml$expr$x_1
yaml$expr$x_1$get_assign

## Example of expression execution //
## The line that contained the YAML markup in 'from' above will be substituted
## by this expression 
x_1 <- 10
eval(yaml$expr$x_1$get_assign)
ref_1

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/fromyaml}
}
\seealso{
\code{
  	\link[fromyaml]{buildExpressionFromYaml-ReactiveReferenceYamlParsed.S3-method}
}
}

