% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{getYaml,function,ReactiveReferenceYamlContext.S3-method}
\alias{getYaml,function,ReactiveReferenceYamlContext.S3-method}
\alias{getYaml-function-ReactiveReferenceYamlContext.S3-method}
\title{Get YAML Markup (function-ReactiveReferenceYamlContext.S3)}
\usage{
\S4method{getYaml}{`function`,ReactiveReferenceYamlContext.S3}(from,
  ctx = NULL, ...)
}
\arguments{
\item{from}{\code{\link{missing}}.}

\item{ctx}{\strong{Signature argument}.
Markup proecessing context.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
\code{\link{character}}. Identified YAML markup.
}
\description{
See generic: \code{\link[fromyaml]{getYaml}}
}
\examples{
\dontrun{

from <- function() {
  "reactive-ref: {id: x_1, where: .GlobalEnv, as: ref_1}"
}
yaml <- getYaml(
  from = from,
  ctx = ReactiveReferenceYamlContext.S3()
)
yaml
yaml$yaml
yaml$index
yaml$src

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/fromyaml}
}
\seealso{
\code{
   \link[fromyaml]{getYaml}
}
}

