% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getYaml}
\alias{getYaml}
\title{Get YAML Markup (generic)}
\usage{
getYaml(from, ctx = NULL, where = parent.frame(), strict = FALSE, ...)
}
\arguments{
\item{from}{\strong{Signature argument}.
Object containing YAML markup.}

\item{ctx}{\strong{Signature argument}.
Markup proecessing context.}

\item{where}{\code{\link{environment}}.
Environment in which to assign the function in the \code{src} field of
class \code{\link[yamlr]{ObjectReferenceYaml.S3}}. Only relevant
in case the YAML has been provided via comments instead of an
inline string as this involves some additional transformation steps.}

\item{strict}{\code{\link{logical}}.
\code{TRUE}: error if no YAML markup could be identified (which in turn
results in field \code{original} in class
\code{\link[yamlr]{ObjectReferencedYaml.S3}} being empty);
\code{FALSE}: no error if no YAML markup could be identified.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\description{
Identifies lines that correspond to YAML markup.
}
\section{YAML syntax}{


Note that the YAML statement for specifying object references must
be as follows:

\code{object-ref: {id: {id}, where: {where}, as: {id-ref}}}

Note that elements \code{where} and \code{as} are optional and that spaces
are required between each syntax component. While the lack of spaces will
cause a parsing error in \code{\link[yaml]{yaml.load}}, additional spaces
do not casue any trouble.

Example for valid YAML statements

\itemize{
   \item{\code{object-ref: {id: x_1, where: .GlobalEnv, as: ref_1}}}
   \item{\code{object-ref: {id: x_1, where: .GlobalEnv}}}
   \item{\code{object-ref: {id: x_1}}}
   \item{\code{object-ref:  {  id:   x_1, where  : .GlobalEnv, as: ref_1  }}}
}

See the official \href{YAML website}{http://www.yaml.org/} for more
details about YAML
}
\examples{
\dontrun{

##------------------------------------------------------------------------------  
## As inline string //
##------------------------------------------------------------------------------

## With curly brackets //  
yaml <- getYaml(
  from = function() {
    "object-ref: {id: x_1, where: .GlobalEnv, as: ref_1}"
    ref_1 * 2
  },
  ctx = YamlContext.ObjectReference.S3()
)
yaml
yaml$original
yaml$index
yaml$src

## W/o curly brackets //
yaml <- getYaml(
  from = function()
    "object-ref: {id: x_1, where: .GlobalEnv, as: ref_1}",
  ctx = YamlContext.ObjectReference.S3()
)
yaml$original
yaml$original
yaml$index
yaml$src

## Expression //
yaml <- getYaml(
  from = captureExpression({
    "object-ref: {id: x_1, where: .GlobalEnv, as: ref_1}"
    ref_1 * 2
  }),
  ctx = YamlContext.ObjectReference.S3()
)
yaml
yaml$original
yaml$index
yaml$src

## Condition handling //
yaml <- getYaml(
  from = function (){
    "Invalid YAML markup"
    ref_1 * 2
  },
  ctx = YamlContext.ObjectReference.S3()
)
yaml
yaml$original
yaml$index
yaml$src

try(getYaml(
  from = function (){
    "Invalid YAML markup"
    ref_1 * 2
  },
  ctx = YamlContext.ObjectReference.S3(),
  strict = TRUE
))

##------------------------------------------------------------------------------  
## As comment //
##------------------------------------------------------------------------------

## With curly brackets //  
yaml <- getYaml(
  from = function() {
    ## object-ref: {id: x_1, where: .GlobalEnv, as: ref_1}
    ref_1 * 2
  },
  ctx = YamlContext.ObjectReference.S3()
)
yaml$original
yaml$index
yaml$src

## W/o curly brackets //
yaml <- getYaml(
  from = function()
    ## object-ref: {id: x_1, where: .GlobalEnv, as: ref_1}
    ref_1 * 2,
  ctx = YamlContext.ObjectReference.S3()
)
yaml$original
yaml$index
yaml$src

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/yamlr}
}
\seealso{
\code{
  	\link[yamlr]{getYaml-function-YamlContext.ObjectReference.S3-method}
}
}

