% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{parseYaml,ReactiveReferenceYaml.S3-method}
\alias{parseYaml,ReactiveReferenceYaml.S3-method}
\alias{parseYaml-ReactiveReferenceYaml.S3-method}
\title{Parse YAML Markup (ReactiveReferenceYaml.S3)}
\usage{
\S4method{parseYaml}{ReactiveReferenceYaml.S3}(yaml, ...)
}
\arguments{
\item{yaml}{\code{\link{ReactiveReferenceYaml.S3}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
\code{\link{ReactiveReferenceYamlParsed.S3}}.
   Parsed YAML markup.
}
\description{
See generic: \code{\link[yamlyaml]{parseYaml}}
}
\examples{
\dontrun{

## Get //
yaml <- getYaml(
  from = function() {
  "reactive-ref: {id: x_1, where: .GlobalEnv, as: ref_1}"
  },
  ctx = YamlContext.ReactiveReference.S3()
)

## Parse //
yaml <- parseYaml(yaml = yaml)

## Inspeact //
yaml
yaml$original
yaml$parsed
yaml$index
yaml$src

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/yamlr}
}
\seealso{
\code{
   \link[yamlyaml]{parseYaml}
}
}

