% Generated by roxygen2 (4.0.2): do not edit by hand
\name{parseYaml}
\alias{parseYaml}
\title{Parse YAML Markup (generic)}
\usage{
parseYaml(yaml, ...)
}
\arguments{
\item{yaml}{\strong{Signature argument}.
Object containing identified YAML markup as returned by
\code{\link[yamlr]{getYaml}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\description{
Parses YAML markup as identified by \code{\link[yamlr]{getYaml}}.
}
\section{YAML syntax}{


Note that the YAML statement for specifying object references must
be as follows:

\code{object-ref: {id: {id}, where: {where}, as: {id-ref}}}

Note that elements \code{where} and \code{as} are optional and that spaces
are required between each syntax component. While the lack of spaces will
cause a parsing error in \code{\link[yaml]{yaml.load}}, additional spaces
do not casue any trouble.

Example for valid YAML statements

\itemize{
   \item{\code{object-ref: {id: x_1, where: .GlobalEnv, as: ref_1}}}
   \item{\code{object-ref: {id: x_1, where: .GlobalEnv}}}
   \item{\code{object-ref: {id: x_1}}}
   \item{\code{object-ref:  {  id:   x_1, where  : .GlobalEnv, as: ref_1  }}}
}

See the official \href{YAML website}{http://www.yaml.org/} for more
details about YAML
}
\examples{
\dontrun{

## Get //
yaml <- getYaml(
  from = function() {
  "object-ref: {id: x_1, where: .GlobalEnv, as: ref_1}"
  },
  ctx = YamlContext.ObjectReference.S3()
)

## Parse //
yaml <- parseYaml(yaml = yaml)

## Inspeact //
yaml
yaml$original
yaml$parsed
yaml$index
yaml$src

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/yamlr}
}
\seealso{
\code{
  	\link[yamlr]{parseYaml-function-YamlContext.ObjectReference.S3-method}
}
}

