% Generated by roxygen2 (4.0.2): do not edit by hand
\name{processYaml}
\alias{processYaml}
\title{Process YAML Markup (generic)}
\usage{
processYaml(from, ctx = NULL, where = parent.frame(), strict = FALSE, ...)
}
\arguments{
\item{from}{\strong{Signature argument}.
Object containing YAML markup.}

\item{ctx}{\strong{Signature argument}.
Markup proecessing context.}

\item{where}{\code{\link{environment}}.
Environment in which to assign the function in the \code{src} field of
class \code{\link[yamlr]{ObjectReferenceYaml.S3}}. Only relevant
in case the YAML has been provided via comments instead of an
inline string as this involves some additional transformation steps.}

\item{strict}{\code{\link{logical}}.
\code{TRUE}: error if no YAML markup could be identified (which in turn
results in field \code{original} in class
\code{\link[yamlr]{ObjectReferencedYaml.S3}} being empty);
\code{FALSE}: no error if no YAML markup could be identified.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\description{
Convenience wrapper. Processes YAML markup in the sense that the complete
workflow associated to a given application context is executed.
}
\details{
Currently, this means that all processing steps associated with the context
\code{\link[yamlr]{YamlContext.ObjectReference.S3}} are executed:

\itemize{
   \item{\code{\link[yamlr]{getYaml}}}
   \item{\code{\link[yamlr]{parseYaml}}}
   \item{\code{\link[yamlr]{buildExpressionFromYaml}}}
   \item{\code{\link[yamlr]{updateYamlSource}}}
}
}
\section{YAML syntax}{


Note that the YAML statement for specifying object references must
be as follows:

\code{object-ref: {id: {id}, where: {where}, as: {id-ref}}}

Note that elements \code{where} and \code{as} are optional and that spaces
are required between each syntax component. While the lack of spaces will
cause a parsing error in \code{\link[yaml]{yaml.load}}, additional spaces
do not casue any trouble.

Example for valid YAML statements

\itemize{
   \item{\code{object-ref: {id: x_1, where: .GlobalEnv, as: ref_1}}}
   \item{\code{object-ref: {id: x_1, where: .GlobalEnv}}}
   \item{\code{object-ref: {id: x_1}}}
   \item{\code{object-ref:  {  id:   x_1, where  : .GlobalEnv, as: ref_1  }}}
}

See the official \href{YAML website}{http://www.yaml.org/} for more
details about YAML
}
\examples{
\dontrun{
  
##------------------------------------------------------------------------------
## From function //
##------------------------------------------------------------------------------

## Get //  
yaml <- processYaml(
  from = function() {
    "object-ref: {id: x_1, where: .GlobalEnv, as: ref_1}"
    ref_1 * 2
  },
  ctx = YamlContext.ObjectReference.S3()
)

## Inspect //
yaml$src
## --> note that the original YAML markup has been substituted by a 
## suitable expression for this specific context (`yaml$src$x_1$get_assign`).
## The processing is now complete and the source object can be used as is
## as it is now "self contained" in the sense that it contains all information
## that are needed in order for it to do its job.

## Actual use of the source object //
x_1 <- 10
eval(yaml$src())

## Clean up //
rm(x_1)
rm(yaml)

##------------------------------------------------------------------------------
## From expression //
##------------------------------------------------------------------------------

## Get //  
yaml <- processYaml(
  from = captureExpression({
    "object-ref: {id: x_1, where: .GlobalEnv, as: ref_1}"
    ref_1 * 2
  }),
  ctx = YamlContext.ObjectReference.S3()
)

## Inspect //
yaml$src

## Actual use of the source object //
x_1 <- 10
eval(yaml$src)

## Clean up //
rm(x_1)
rm(yaml)

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/yamlr}
}
\seealso{
\code{
  	\link[yamlr]{processYaml-function-YamlContext.ObjectReference.S3-method},
   \link[yamlr]{getYaml},
   \link[yamlr]{parseYaml},
   \link[yamlr]{buildExpressionFromYaml},
   \link[yamlr]{updateYamlSource}
}
}

