% Generated by roxygen2 (4.0.2): do not edit by hand
\name{updateYamlSource}
\alias{updateYamlSource}
\title{Update Object that originally contained YAML Markup (generic)}
\usage{
updateYamlSource(yaml, ...)
}
\arguments{
\item{yaml}{\strong{Signature argument}.
Object containing processed YAML markup as returned by
\code{\link[fromyaml]{buildExpressionFromYaml}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\description{
Updates the object that originally contained YAML markup. To be more
specific, the function updates field \code{src} or class
\code{\link[fromyaml]{ReactiveReferenceYamlProcessed}} as returned by
\code{\link[fromyaml]{buildExpressionFromYaml}}.
}
\examples{
\dontrun{

## Get //  
yaml <- getYaml(
  from = function() {
  "reactive-ref: {id: x_1, where: .GlobalEnv, as: ref_1}"
  ref_1 * 2
  },
  ctx = ReactiveReferenceYamlContext.S3()
)

## Parse //
yaml <- parseYaml(yaml = yaml)

## Build //
yaml <- buildExpressionFromYaml(yaml = yaml)

## Update //
yaml <- updateYamlSource(yaml = yaml)

## Inspect //
yaml$src
## --> note that the original YAML markup has been substituted by a 
## suitable expression for this specific context (`yaml$src$x_1$get_assign`).
## The processing is now complete and the source object can be used as is
## as it is now "self contained" in the sense that it contains all information
## that are needed in order for it to do its job.

## Actual use of the source object //
x_1 <- 10
eval(yaml$src())

## Clean up //
rm(x_1)
rm(yaml)
rm(yaml)

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/fromyaml}
}
\seealso{
\code{
  	\link[fromyaml]{updateYamlSource-ReactiveReferenceYamlProcessed.S3-method},
    \link[fromyaml]{getYaml},
    \link[fromyaml]{parseYaml},
    \link[fromyaml]{buildExpressionFromYaml}
}
}

